+++ src/database/migrations/2003_07_29_000002_create_lwc_admin_zones_table.php
@@ -3,7 +3,7 @@
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 
-class CreateLwcCitiesTable extends Migration
+class CreateLwcAdminZonesTable extends Migration
 {
 
     /**
@@ -13,14 +13,15 @@ class CreateLwcCitiesTable extends Migration
      */
     public function up()
     {
-        Schema::create('lwc_cities', function (Blueprint $table) {
+        Schema::create('lwc_admin_zones', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('name', 200);
+            $table->enum('type', ['adm1', 'adm2', 'adm3', 'adm4']);
+            $table->string('adm1', 20);
+            $table->string('adm2', 80);
+            $table->string('adm3', 20);
+            $table->string('adm4', 20);
             $table->string('country_iso2', 2);
-            $table->string('admin1', 20);
-            $table->string('admin2', 80);
-            $table->string('admin3', 20);
-            $table->string('admin4', 20);
             $table->decimal('latitude', 10, 7);
             $table->decimal('longitude', 10, 7);
         });
@@ -34,6 +35,6 @@ class CreateLwcCitiesTable extends Migration
      */
     public function down()
     {
-        Schema::drop('lwc_cities');
+        Schema::drop('lwc_admin_zones');
     }
 }