+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v2.0.0
+
+- Add administrative subdivisions into `lwc_admin_zones` table
+
 # v1.0.0
 
-First version : adds lwc_cities table and populates it with cities.
+First version : adds `lwc_cities` table and populates it with cities.